% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_bis.R
\name{json_bis}
\alias{json_bis}
\title{json_bis: Generates a json file for the Body Image scale}
\usage{
json_bis()
}
\value{
A string with data stored in JSON format containing meta-data for the Body Image scale
}
\description{
This function generates a json file for the scored Body Image scale and raw participant responses.
}
