% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_food_paradigm.R
\name{json_food_paradigm}
\alias{json_food_paradigm}
\title{json_food_paradigm: Generates a json file for food paradigm notes and data}
\usage{
json_food_paradigm()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for visit 1 food paradigm notes and data
}
