% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_food_paradigm_v5.R
\name{json_food_paradigm_v5}
\alias{json_food_paradigm_v5}
\title{json_food_paradigm_v5: Generates a json file for visit 5 food paradigm notes and data}
\usage{
json_food_paradigm_v5()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for visit 5food paradigm notes and data
}
