% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_fullness.R
\name{json_fullness}
\alias{json_fullness}
\title{json_fullness: Generates a json file for Freddy Fullness data}
\usage{
json_fullness()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for Freddy Fullness data
}
