% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_intake_v5.R
\name{json_intake_v5}
\alias{json_intake_v5}
\title{json_intake_v5: Generates a json file for intake data for visit 5}
\usage{
json_intake_v5()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for intake data for visit 5
}
