% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pst.R
\name{json_pst}
\alias{json_pst}
\title{json_pst: Generates a json file for the child portion size task}
\usage{
json_pst()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for he child portion size task
}
