% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_rsa_v5.R
\name{json_rsa_v5}
\alias{json_rsa_v5}
\title{json_rsa_v5: Generates a json file for RSA during the standard visit elements}
\usage{
json_rsa_v5()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for RSA during the standard visit elements
}
