% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_rsa_visit.R
\name{json_rsa_visit}
\alias{json_rsa_visit}
\title{json_rsa_visit: Generates a json file for RSA during the standard visit elements}
\usage{
json_rsa_visit()
}
\value{
A string with data stored in JSON format containing meta-data
}
\description{
This function generates a json file for RSA during the standard visit elements
}
