% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_social_dq.R
\name{json_social_dq}
\alias{json_social_dq}
\title{json_social_dq: Generates a json file for the Social Desirability Questionnaire}
\usage{
json_social_dq()
}
\value{
A string with data stored in JSON format containing meta-data for the Social Desirability Questionnaire
}
\description{
This function generates a json file for the scored Social Desirability Questionnaire and raw participant responses.
}
