% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_redcap.R
\name{proc_redcap}
\alias{proc_redcap}
\title{proc_redcap: Process raw data downloaded from Study REACH REDCap}
\usage{
proc_redcap(redcap_api = FALSE, redcap_visit_data, redcap_de_data)
}
\arguments{
\item{redcap_visit_data}{REDCap visit data from a prior API call}
}
\value{
Will return a list including data and metadata for:
#' \itemize{
\item{'paticipants' - BIDS specified participants.tsv file}
\item{'anthropometrics' - height, weight, and computed anthropometric data}
\item{'demographics' - compiled demographic data}
\item{'dxa' - verified DXA data}
\item{'household' - compiled demographicinformation about houshold}
\item{'infancy' - compiled demographic information related to infancy}
\item{'intake' - compiled intake data with computed intake values}
\item{'parent_updates' - all visit updates}
\item{'researcher_notes' - all visit notes}
\item{'asbi' - Adaptive Social Behavioral Inventory}
\item{'bevqps' - Beverage Questionnaire for Preschoolers}
\item{'brief2' - Behavioral Rating Inventory of Executive Function}
\item{'cbq' - Child Behavior Questionnaire}
\item{'cebq' - Children's Eating Behavior Questionnaire}
\item{'cfsq' - Caregiver's Feeding Style Questionnaire}
\item{'dor' - Division of Responsibility}
\item{'ecsi2' - Parent Eating Competence Questionnaire}
\item{'tfeq' - Three Factor Eating Questionnaire}
\item{'nns_ffq' - Non-Nutritive Sweetener FFQ (2 versions - parent and child)}
\item{'pwlb' - Parent Weight Loss Behaviors}
}
}
\description{
This function:
\itemize{
\item{1) Reads REDCap data (visit and double-entry) using the REDCap API}
\item{2) Calls util_ functions to clean and compile data in dataframes}
\item{3) Calls json_ functions to create strings with meta-data stored in JSON format for each dataframe}
\item{4) Compiles data repeated across visits and sessions}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
redcap_data <- proc_redcap(base_wd, overwrite = FALSE, overwrite_jsons = FALSE)

}

}
\seealso{
\code{\link[=write_redcap]{write_redcap()}}
}
