% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_cfq_data.R
\name{util_format_cfq_data}
\alias{util_format_cfq_data}
\title{util_format_cfq_data: clean up the administration of the Child Feeding Questionnaire}
\usage{
util_format_cfq_data(cfq_data)
}
\arguments{
\item{cfq_data}{rank extracted from data from REDCap events}
}
\description{
This function cleans up the administration of the Child Feeding Questionnaire (skipped questions about child weight concerns so numbers are off)
}
\examples{

# process data
cfq_data_formatted <- util_format_cfq_data(cfq_data)

}
\seealso{
\code{\link[=util_redcap_parent_v1]{util_redcap_parent_v1()}}
}
