% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_infant_data.R
\name{util_format_infant_data}
\alias{util_format_infant_data}
\title{util_format_infant_data: process infant birth weight data}
\usage{
util_format_infant_data(infancy_data)
}
\arguments{
\item{infancy_data}{infancy data extracted from data from REDCap events}
}
\description{
This function process nfant birth weight data
}
\examples{

# process data
infancy_data_formatted <- util_format_infant_data(infancy_data)

}
\seealso{
\code{\link[=util_redcap_child_v1]{util_redcap_child_v1()}}
}
