% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_intake_data.R
\name{util_format_intake_data}
\alias{util_format_intake_data}
\title{util_format_intake_data: process preload and meal intake data}
\usage{
util_format_intake_data(intake_data)
}
\arguments{
\item{intake_data}{intake data extracted from data from REDCap events}
}
\description{
This function process Ranking Food Item Questionnaire data
}
\examples{

# process data
intake_data_formatted <- util_format_intake_data(intake_data)

}
\seealso{
\code{\link[=util_redcap_child_v1]{util_redcap_child_v1()}}, \code{\link[=util_redcap_child_v2]{util_redcap_child_v2()}}, \code{\link[=util_redcap_child_v3]{util_redcap_child_v3()}}, \code{\link[=util_redcap_child_v4]{util_redcap_child_v4()}}, \code{\link[=util_redcap_child_v5]{util_redcap_child_v5()}}
}
