% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent_v4.R
\name{util_redcap_parent_v4}
\alias{util_redcap_parent_v4}
\title{util_redcap_parent_v4: Organize parent visit 4 data from REDCap}
\usage{
util_redcap_parent_v4(data, date_data)
}
\arguments{
\item{data}{data from REDCap event parent_visit_4_arm_1}

\item{date_data}{date data for REDCap visit

Will return a list including:
\itemize{
\item{clean raw and scored parent visit 4 datasets}
\item{meta-data formated as json for each dataset}
}

Returned data includes:
\itemize{
\item{ecsi_data}
\item{cfsq_data}
\item{asbi_data}
}}
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_4_arm_1
}
\examples{

# process REDCap data
parent_visit4_list <- util_redcap_parent_v4(data, date_data)

\dontrun{
}

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}
}
