% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent_v5.R
\name{util_redcap_parent_v5}
\alias{util_redcap_parent_v5}
\title{util_redcap_parent_v5: Organize parent visit 5 data from REDCap}
\usage{
util_redcap_parent_v5(data, date_data)
}
\arguments{
\item{data}{data from REDCap event parent_visit_5_arm_1}

\item{date_data}{date data for REDCap visit

Will return a list including:
\itemize{
\item{clean raw and scored parent visit 5 datasets}
\item{meta-data formated as json for each dataset}
}

Returned data includes:
\itemize{
\item{nnsa_data}
\item{nns_parent_data}
\item{nns_child_data}
\item{bevq_data}
}}
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_5_arm_1
}
\examples{

# process REDCap data
parent_visit5_list <- util_redcap_parent_v5(data, date_data)

\dontrun{
}

}
\seealso{
\code{\link[=proc_redcap]{proc_redcap()}}
}
